Refactor web session responsibilities into dedicated mixins
